create reducers for state, move functions in utils
