+++ .eslintrc.json
@@ -26,6 +26,7 @@
         "prefer-rest-params": 0,
         "react/display-name": 0,
         "@typescript-eslint/no-unused-vars": 0,
-        "@typescript-eslint/ban-ts-comment": 0
+        "@typescript-eslint/ban-ts-comment": 0,
+        "@typescript-eslint/no-this-alias": 1
     }
 }
