+++ package.json
@@ -28,6 +28,7 @@
     "react-router": "^5.2.0",
     "react-select": "^4.0.2",
     "redux": "^4.0.5",
+    "redux-actions": "^2.6.5",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
