+++ src/actions/types.ts
@@ -26,3 +26,6 @@ export const ADD_BOOKING_ERRORS = 'ADD_BOOKING_ERRORS';
 export const DELETE_BOOKING = 'DELETE_BOOKING';
 export const DELETE_BOOKING_SUCCESS = 'DELETE_BOOKING_SUCCESS';
 export const DELETE_BOOKING_ERRORS = 'DELETE_BOOKING_ERRORS';
+
+export const UPDATED_SELECTED = 'UPDATED_SELECTED';
+export const UPDATED_CURRENT_BOOKING = 'UPDATED_CURRENT_BOOKING';
