+++ src/reducers/index.ts
@@ -3,6 +3,8 @@ import { connectRouter } from 'connected-react-router';
 import errorReducer from './errors-reducers';
 import {rooms, roomsHasErrored, roomsIsLoading, TypeRoom} from './rooms-reducers';
 import {booking, bookingHasErrored, bookingIsLoading, TypeBooking} from './booking-reducers';
+import {selected, TypeSelected} from './selected-reducers';
+import {currentBooking} from './current-booking-reducers';
 
 const rootReducer = (history) => combineReducers({
     router: connectRouter(history),
@@ -12,18 +14,22 @@ const rootReducer = (history) => combineReducers({
     roomsIsLoading,
     booking,
     bookingHasErrored,
-    bookingIsLoading
+    bookingIsLoading,
+    selected,
+    currentBooking
 })
 
 export default rootReducer;
 
 export interface RootState {
-    router: Function;
+    router: any;
     errors: any;
     rooms: TypeRoom[];
     roomsHasErrored: boolean;
     roomsIsLoading: boolean;
     booking: TypeBooking[];
+    selected: TypeSelected;
     bookingHasErrored: boolean;
     bookingIsLoading: boolean;
+    currentBooking: TypeRoom;
 }
