+++ src/views/Booking.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
 import {useDispatch, shallowEqual, useSelector} from 'react-redux';
 import {itemsRoomsFetchData} from '../actions/rooms-actions';
 import {clearErrors, deleteBooking, createBooking, itemsBookingFetchData, updateBooking} from '../actions/booking-actions';
@@ -11,7 +11,8 @@ import {TypeRoom} from '../reducers/rooms-reducers';
 import {RootState} from '../reducers';
 import {TypeBooking} from '../reducers/booking-reducers';
 import moment, {Moment} from 'moment';
-import {isGoodRange, timenull} from "../helpers/booking-helpers";
+import {isGoodRange, timenull, updateSelected, setCurrentBooking, changeBooking} from "../utils/booking-utils";
+import {TypeSelected} from "../reducers/selected-reducers";
 
 export interface IContextProps {
     openWindow: boolean;
@@ -19,23 +20,21 @@ export interface IContextProps {
     currentBooking: TypeBooking;
 }
 
-interface TypeItemSelected {
-    day: null | Moment;
-    room: string | boolean;
-}
-
-interface TypeSelected {
-    start: TypeItemSelected;
-    end: TypeItemSelected;
-}
-
 export const WindowContext = React.createContext({});
 
 export const Booking = () => {
+
     const bookingIsLoading: boolean = useSelector((state: RootState) => state.bookingIsLoading, shallowEqual);
     const bookingHasErrored: (a: boolean) => boolean = useSelector((state: RootState) => state.bookingHasErrored, shallowEqual);
     const [openWindow, setOpenWindow] = useState(false);
-    const [currentBooking, setCurrentBooking] = useState({
+    const selected: TypeSelected = useSelector((state: RootState) => state.selected, shallowEqual);
+    const currentBooking: TypeBooking = useSelector((state: RootState) => state.currentBooking, shallowEqual);
+
+    const rooms: TypeRoom[] = useSelector((state: RootState) => state.rooms, shallowEqual)
+    const booking: TypeBooking[] = useSelector((state: RootState) => state.booking, shallowEqual)
+    const dispatch = useDispatch()
+    /*
+     const [currentBooking, setCurrentBooking] = useState({
         _id: 'create',
         fio: '',
         cost: '',
@@ -43,13 +42,12 @@ export const Booking = () => {
         startDate: moment(),
         endDate: moment()
     });
-    const rooms: TypeRoom[] = useSelector((state: RootState) => state.rooms, shallowEqual)
-    const booking: TypeBooking[] = useSelector((state: RootState) => state.booking, shallowEqual)
-    const dispatch = useDispatch()
     const [selected, setSelected] = useState({
         start:{day:null, room:false},
         end:{day:null, room:false}
     })
+
+     */
 
     useEffect(() => {
         if(rooms.length == 0)
@@ -58,6 +56,20 @@ export const Booking = () => {
             dispatch(itemsBookingFetchData('/api/booking/getall'))
     }, []);
 
+    const actionClose = () => {
+        dispatch(updateSelected({
+            start:{day:null, room:false},
+            end:{day:null, room:false}
+        }))
+        /*
+        setSelected({
+            start:{day:null, room:false},
+            end:{day:null, room:false}
+        });
+
+         */
+    }
+    /*
     useEffect(() => {
         if(!openWindow) {
             setSelected({
@@ -66,7 +78,7 @@ export const Booking = () => {
             });
         }
     }, [openWindow])
-
+*/
     document.title = "Booking"
 
     if (bookingHasErrored) {
@@ -110,120 +122,32 @@ export const Booking = () => {
         }
     }
 
-    const onClickBooking = (id) => {
+    const onClickBooking = (id: string) => {
         dispatch(clearErrors())
         const editBooking = booking.filter(item => item._id == id)[0]
-        setCurrentBooking(editBooking)
+        dispatch(setCurrentBooking(editBooking))
         setOpenWindow(true)
     }
 
-    const changeBooking = (startDate, endDate, room = '') => {
-        if ((selected.start.day && selected.end.day && currentBooking._id == 'create') || currentBooking._id != 'create') {
-            if(currentBooking._id == 'create') {
-                if(room == '')
-                    room = selected.start.room;
-                if(moment(timenull(startDate)).isSame(timenull(endDate))) {
-                    alert('Дата выезда и дата въезда должны различаться.')
-                    return [
-                        selected.start.day.toDate(),
-                        selected.end.day.toDate(),
-                        selected.start.room
-                    ]
-                }
-                else if(moment(startDate).diff(endDate, 'days') < 0 &&  isGoodRange(moment(startDate), moment(endDate),  room)) {
-                    setSelected({
-                        start: {day: moment(timenull(startDate)), room: room},
-                        end: {day: moment(timenull(endDate)), room: room}
-                    });
-                    return true
-                }
-                else if(moment(timenull(startDate)).diff(timenull(endDate), 'days') > 0 && isGoodRange(moment(endDate), moment(startDate), room)) {
-                    setSelected({
-                        start: {day: moment(timenull(endDate)), room: room},
-                        end: {day: moment(timenull(startDate)), room: room}
-                    })
-                    return [
-                        endDate,
-                        startDate,
-                        selected.start.room
-                    ]
-                }
-                else {
-                    alert('На этот дипазон уже назначено бронирование.')
-                    return [
-                        new Date(selected.start.day.format('YYYY-MM-DD')),
-                        new Date(selected.end.day.format('YYYY-MM-DD')),
-                        selected.start.room
-                    ]
-                }
-            }
-            else {
-                if(room == '')
-                    room = currentBooking.room;
-                if(moment(timenull(startDate)).isSame(timenull(endDate))) {
-                    alert('Дата выезда и дата въезда должны различаться.')
-                    return [
-                        new Date(currentBooking.startDate),
-                        new Date(currentBooking.endDate),
-                        currentBooking.room
-                    ]
-                }
-                else if(moment(startDate).diff(endDate, 'days') < 0 && isGoodRange(moment(startDate), moment(endDate), room, currentBooking._id)) {
-                    const curRoom = rooms.filter((item) => item._id == room)[0];
-                    const cost = Math.abs(moment(startDate).diff(moment(endDate), "days") * parseInt(curRoom.cost));
-                    setCurrentBooking({
-                        _id: currentBooking._id,
-                        fio: currentBooking.fio,
-                        cost: cost,
-                        room: room,
-                        startDate: moment(startDate).format('YYYY-MM-DD'),
-                        endDate: moment(endDate).format('YYYY-MM-DD')
-                    })
-                    return true
-                }
-                else if(moment(startDate).diff(endDate, 'days') > 0 && isGoodRange(moment(endDate), moment(startDate), room, currentBooking._id)) {
-                    const curRoom = rooms.filter((item) => item._id == room)[0];
-                    const cost = Math.abs(moment(endDate).diff(moment(startDate), "days") * parseInt(curRoom.cost));
-                    setCurrentBooking({
-                        _id: currentBooking._id,
-                        fio: currentBooking.fio,
-                        cost: cost,
-                        room: room,
-                        startDate: moment(endDate).format('YYYY-MM-DD'),
-                        endDate: moment(startDate).format('YYYY-MM-DD')
-                    })
-                    return [
-                        moment(endDate).toDate(),
-                        moment(startDate).toDate(),
-                        currentBooking.room
-                    ]
-                }
-                else {
-                    alert('На этот дипазон уже назначено бронирование.')
-                    return [
-                        new Date(currentBooking.startDate),
-                        new Date(currentBooking.endDate),
-                        currentBooking.room
-                    ]
-                }
-            }
+    const changeSelected = (newSelected: TypeSelected) => {
+
+        const curRoom = rooms.filter((item) => item._id == newSelected.start.room)[0]
+        if(typeof curRoom != 'undefined') {
+            const cost = Math.abs(moment(newSelected.start.day).diff(moment(newSelected.end.day), "days") * parseInt(curRoom.cost))
+            dispatch(setCurrentBooking({
+                _id: 'create',
+                cost: cost + '',
+                fio: '',
+                room: newSelected.start.room,
+                startDate: newSelected.start.day.toDate(),
+                endDate: moment(newSelected.end.day).toDate()
+            }))
+            dispatch(updateSelected(newSelected))
+            //setSelected(newSelected)
+            dispatch(clearErrors())
+            setOpenWindow(true)
+
         }
-    }
-
-    const changeSelected = (selected) => {
-        const curRoom = rooms.filter((item) => item._id == selected.start.room)[0]
-        const cost = Math.abs(moment(selected.start.day).diff(moment(selected.end.day), "days") * parseInt(curRoom.cost))
-        setCurrentBooking({
-            _id: 'create',
-            cost: cost+'',
-            fio: '',
-            room: selected.start.room,
-            startDate: selected.start.day.format('YYYY-MM-DD'),
-            endDate: moment(selected.end.day).format('YYYY-MM-DD')
-        })
-        setSelected(selected)
-        dispatch(clearErrors())
-        setOpenWindow(true)
     }
 
     return  <div id="wrapper">
@@ -237,12 +161,14 @@ export const Booking = () => {
                     startDate={timenull(moment())}
                     endDate={timenull(moment().add(13,'days'))}
                 />
-                <WindowContext.Provider value={{ openWindow, setOpenWindow, currentBooking}}>
+                <WindowContext.Provider value={{ openWindow, setOpenWindow }}>
                     <BookingModal
                         onChangeBooking={changeBooking}
                         onActionModal={actionModal}
                         onActionDelete={actionDelete}
+                        onActionClose={actionClose}
                     />
+                    <useClickBooking/>
                 </WindowContext.Provider>
             </div>
 }
