+++ src/views/BookingModal.tsx
@@ -13,11 +13,12 @@ import "react-datepicker/dist/react-datepicker.css";
 import Select from 'react-select';
 import {EventBus} from "../events";
 import {RootState} from "../reducers";
+import {TypeBooking} from "../reducers/booking-reducers";
 
 const modalStyles = {
     overlay: { zIndex: 100 },
     content : {top : '50%', left : '50%', right : 'auto', bottom : 'auto', marginRight : '-50%', boxShadow : '0 0 10px rgba(0,0,0,0.5)', width : '500px', transform  : 'translate(-50%, -50%)'}
-};
+}
 
 const customStyles = {
     control: (styles, isFocused) => ({ ...styles, backgroundColor: 'white', boxShadow: isFocused ? '' : '', border: isFocused ? '0px solid grey' : '0px solid grey' }),
@@ -35,37 +36,32 @@ type ModalProps = {
     onActionDelete: (id: string) => void;
     onActionClose?: () => void;
     onActionModal: (cost: string, fio: string, room: string, startDate: Moment | Date, endDate: Moment | Date) => void;
-};
+}
 
 export const BookingModal = (props: ModalProps) => {
-    const errors: any = useSelector((state: RootState) => state.errors, shallowEqual);
-    const rooms: TypeRoom[] = useSelector((state: RootState) => state.rooms, shallowEqual);
+    const errors: any = useSelector((state: RootState) => state.errors, shallowEqual)
+    const rooms: TypeRoom[] = useSelector((state: RootState) => state.rooms, shallowEqual)
+    const currentBooking: TypeBooking = useSelector((state: RootState) => state.currentBooking, shallowEqual)
 
-    const [fio, setFio] = useState('');
-    const [startDate, setStartDate] = useState(new Date());
-    const [endDate, setEndDate] = useState(new Date());
-    const [cost, setCost] = useState('0');
-    const [room, setRoom] = useState('');
+    const [fio, setFio] = useState('')
+    const [startDate, setStartDate] = useState(new Date())
+    const [endDate, setEndDate] = useState(new Date())
+    const [cost, setCost] = useState('0')
+    const [room, setRoom] = useState('')
 
-    const [roomsOptions, setRoomsOptions] = useState([]);
-    const [defRoomsOptions, setDefRoomsOptions] = useState([]);
+    const [roomsOptions, setRoomsOptions] = useState([])
+    const [defRoomsOptions, setDefRoomsOptions] = useState([])
 
-    const { openWindow, setOpenWindow, currentBooking } = useContext(WindowContext);
+    const { openWindow, setOpenWindow } = useContext(WindowContext)
     EventBus.subscribe('bookingCloseWindow', () => closeWindow())
 
     useMemo(() => {
-        const roomsOptions: TypeRoomsOptions[] = [];
+        const roomsOptions: TypeRoomsOptions[] = []
         rooms.map((room) => {
             roomsOptions.push({label: room.num, value: room._id})
         });
-        setRoomsOptions(roomsOptions);
-    }, [rooms]);
-
-    useEffect(() => {
-        //Modal.setAppElement('#root')
-        //Modal.setAppElement('*');
-        //Modal.setAppElement('body');
-    }, []);
+        setRoomsOptions(roomsOptions)
+    }, [rooms])
 
     useEffect(() => {
         if(openWindow) {
@@ -79,17 +75,17 @@ export const BookingModal = (props: ModalProps) => {
             setCost(currentBooking.cost)
             setRoom(currentBooking.room)
         }
-    }, [openWindow]);
+    }, [openWindow])
 
     const closeWindow = () => {
         if(setOpenWindow)
-            setOpenWindow(false);
+            setOpenWindow(false)
         if(props.onActionClose)
-            props.onActionClose();
+            props.onActionClose()
     }
 
     const changeBooking = (newStartDate, newEndDate, room) => {
-        let result = props.onChangeBooking(newStartDate, newEndDate, room);
+        const result = props.onChangeBooking(newStartDate, newEndDate, room)
         if(typeof result != 'undefined') {
             if (result === true) {
                 const curRoom = rooms.filter((item) => item._id == room)[0];
@@ -100,11 +96,11 @@ export const BookingModal = (props: ModalProps) => {
                     label: curRoom.num,
                     value: room
                 })
-                const cost = Math.abs(moment(startDate).diff(moment(endDate), "days") * parseInt(curRoom.cost));
-                setCost(cost+'');
+                const cost = Math.abs(moment(startDate).diff(moment(endDate), "days") * parseInt(curRoom.cost))
+                setCost(cost+'')
             }
             else {
-                const curRoom = rooms.filter((item) => item._id == result[2])[0];
+                const curRoom = rooms.filter((item) => item._id == result[2])[0]
                 setStartDate(result[0])
                 setEndDate(result[1])
                 setRoom(result[2])
@@ -112,8 +108,8 @@ export const BookingModal = (props: ModalProps) => {
                     label: curRoom.num,
                     value: result[2]
                 })
-                const cost = Math.abs(moment(startDate).diff(moment(endDate), "days") * parseInt(curRoom.cost));
-                setCost(cost+'');
+                const cost = Math.abs(moment(startDate).diff(moment(endDate), "days") * parseInt(curRoom.cost))
+                setCost(cost+'')
             }
         }
     }
